requirements.txt
streamlit>=1.30
requests
line-bot-sdk>=3.0
google-analytics-data
google-auth
google-auth-oauthlib
google-api-python-client
apscheduler
pandas
plotly
